\name{GSAfisherCombined-package}
\alias{GSAfisherCombined-package}
\alias{GSAfisherCombined}
\docType{package}
\title{
\packageTitle{GSAfisherCombined}
}
\description{
\packageDescription{GSAfisherCombined}
}
\details{

Package: GSAfisherCombined
Type: Package
Version: 1.0
Date: 2018-02-10
License: GPL (>= 2)
}
\author{
\packageAuthor{GSAfisherCombined}

Maintainer: \packageMaintainer{GSAfisherCombined}
}
\references{
Fisher, R.A. (1925). Statistical Methods for Research Workers. ISBN 0-05-002170-2.
}
